all_requirements.txt
Cython
cytoolz
absl-py==0.4.1
alabaster==0.7.12
allennlp==0.8.0
amqp==2.3.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
aws-xray-sdk==0.95
awscli==1.16.81
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
bleach==2.1.4
bokeh==1.0.3
boto==2.49.0
boto3==1.9.4
botocore==1.12.4
bs4==0.0.1
business-calendar==0.2.1
bz2file==0.98
cached-property==1.5.1
cachetools==2.1.0
camelot-py==0.7.1
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
chartify==2.3.5
cld2-cffi==0.1.4
Click==7.0
client==0.0.1
cloudpickle==0.7.0
colorama==0.3.9
colour==0.1.5
configparser==3.5.0
conllu==0.11
constantly==15.1.0
cookies==2.2.1
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
cymem==1.31.2
cypari==2.3.0
dask==1.1.0
date-extractor==3.9.1
datefinder==0.6.1
decorator==4.3.0
defusedxml==0.5.0
Deprecated==1.2.4
dill==0.2.8.2
docker==3.6.0
docker-pycreds==0.4.0
docutils==0.14
dopamine-rl==2.0.1
dukpy==0.2.2
ecdsa==0.13
editdistance==0.5.2
entrypoints==0.2.3
enum34==1.1.6
ephem==3.7.6.0
et-xmlfile==1.0.1
excalibur-py==0.4.1
Faker==1.0.1
fasttext==0.8.3
flair==0.4.0
flaky==3.4.0
flasgger==0.9.1
Flask==1.0.2
Flask-Cors==3.0.7
ftfy==4.4.3
funcy==1.11
future==0.16.0
fuzzywuzzy==0.17.0
FXrays==1.3.3
gast==0.2.0
gensim==3.4.0
geos==0.2.1
gevent==1.3.6
gin-config==0.1.4
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.5.8
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.15.0
gunicorn==19.9.0
gym==0.11.0
h5py==2.8.0
hdbscan==0.8.18
html5lib==1.0.1
HTMLParser==0.0.2
httplib2==0.12.1
hyperlink==18.0.0
hyperopt==0.1.1
hypothesis==2.0.0
idna==2.7
ijson==2.3
imageio==2.5.0
imagesize==1.1.0
imbalanced-learn==0.3.3
imblearn==0.0
imgaug==0.2.8
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
javascripthon==0.10
jdcal==1.4
jedi==0.12.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
jsondiff==1.1.1
jsonpickle==1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-echarts-pypkg==0.1.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-pip==0.3.1
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications
keras-bert==0.24.0
keras-layer-normalization==0.2.0
keras-multi-head==0.14.0
keras-pos-embd==0.6.0
keras-position-wise-feed-forward==0.1.0
Keras-Preprocessing
keras-self-attention==0.31.0
keras-transformer==0.11.0
keyring==18.0.0
kiwisolver==1.0.1
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
lda==1.1.0
lesscpy==0.13.0
livereload==2.6.0
lml==0.0.2
ludwig==0.1.0
lunardate==0.1.5
lxml==4.2.5
macropy3==1.1.0b2
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mesh-tensorflow==0.0.5
mistune==0.8.3
mkdocs==1.0.4
mock==2.0.0
more-itertools==4.3.0
moto==1.3.4
mpl-finance==0.10.0
mpld3==0.3
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
mysqlclient==1.4.2.post1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.7.2
notedown==1.5.1
numexpr==2.6.9
numpydoc==0.8.0
oauth2client==4.1.3
opencc-python-reimplemented==0.1.4
opencv-python
openpyxl==2.5.9
overrides==1.9
packaging==18.0
pandas==0.23.4
pandoc-attributes==0.1.7
pandocfilters==1.4.2
parsel==1.5.0
parsimonious==0.8.0
parso==0.3.1
pathlib==1.0.1
pbr==3.1.1
pdf2image==1.4.2
pdfminer.six==20181108
pdfminer3k==1.3.1
pickleshare==0.7.4
Pillow==5.2.0
pkuseg==0.0.11
plac==0.9.6
plink==2.2
pluggy==0.7.1
ply==3.11
preshed==1.0.1
prometheus-client==0.3.1
promise==2.2.1
prompt-toolkit==1.0.15
protobuf==3.6.1
py==1.6.0
pyaml==18.11.0
pyarrow==0.12.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyCalverter==1.6.1
pycparser==2.18
pycryptodome==3.6.6
PyDispatcher==2.0.5
pydot==1.4.1
pyecharts==0.2.0
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyemd==0.5.1
pyglet==1.3.2
Pygments==2.2.0
PyHamcrest==1.9.0
pyLDAvis==2.1.2
pylint==2.1.1
pymongo==3.7.2
pymssql
PyMuPDF==1.13.20
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.5
pypiwin32==223
pypng==0.0.18
pyquery==1.4.0
pyreadline==2.1
pytest==3.8.0
python-dateutil==2.7.3
python-jose==2.0.2
python-Levenshtein==0.12.0
PythonMagick==0.9.19
pytorch-pretrained-bert==0.3.0
pytz==2017.3
PyWavelets==1.0.1
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==4.2b1
pyzmq==17.1.2
qtconsole==4.4.1
QtPy==1.5.0
queuelib==1.5.0
regex==2017.4.5
requests==2.20.0
requests-toolbelt==0.9.1
responses==0.10.5
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.2
scikit-surprise==1.0.3
scipy==1.0.0
Scrapy==1.4.0
seaborn==0.9.0
segtok==1.5.7
selenium==3.8.0
Send2Trash==1.5.0
service-identity==17.0.0
shap
Shapely==1.6.4.post1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sklearn==0.0
smart-open==1.6.0
snappy-manifolds==1.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
spacy==2.0.7
spherogram==1.8.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.15
sqlitedict==1.6.0
sqlparse==0.2.4
stop-words==2018.7.23
suds==0.4
sympy==1.3
TA-Lib==0.4.17
tabulate==0.8.3
tensor2tensor
tensorboardX
tensorflow-metadata
tensorflow-probability
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
textacy==0.6.2
textrank4zh==0.3
tf-agents==0.2.0rc2
tfds-nightly==1.0.1.dev201902190105
tflearn==0.3.2
thinc==6.10.2
toolz==0.9.0
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
tuchart==0.1.2.7
tushare
Twisted==17.9.0
typed-ast==1.1.0
typing==3.6.4
ujson==1.35
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
vine==1.2.0
virtualenv==16.0.0
w3lib==1.19.0
Wand==0.4.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
workalendar==2.5.0
wrapt==1.10.11
xgboost
xlrd==1.1.0
xmltodict==0.11.0
zeep==3.3.1
zope.interface==4.5.0
